src/main/java/com/cardstox/cards/controllers/CardsController.java
package com.cardstox.cards.controllers;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RestController;

import com.cardstox.cards.entities.Card;
import com.cardstox.cards.services.CardService;

@RestController
@RequestMapping(value = "/cards")
public class CardsController {

	@Autowired
	CardService cardService;

	@RequestMapping(method = RequestMethod.GET)
	public List<Card> findAllCards() {
		return cardService.findAllCards();
	}
}
